requirements.txt
pillow
flask
pykakasi
numpy
opencv-python